Fix empty trees erroring out
We had added a check for empty trees during pull to make sure that if we misparse
the tree, we error out rather than send an invalid repo.
This turned out to be wrong: empty Git trees do exist.
This patch fixes this by making sure that if the tree object is actually empty
(objsize 0), and no entries were found, it is considered okay.

Fixes: #26
Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>