Fix state where announce is blocking with lock held
This fixes the case where an announcement happens on a repo right as
another push is trying to unsubscribe: the channels are unbuffered,
which means that the announcement waits for the messages to be submitted,
but nothing is processing them anymore.

Make sure that while we wait on the lock, all submitted messages get
consumed.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>