Allow overriding bridge CA and baseurl from environment
The bridge needs to use as little implicit behavior as possible, as such
we don't want to automatically load everything from environment unless the
calling application says it will take full control.
However, it seems people want to be able to override just the BaseURL
from environment, so let's add a special :FROMENV: configuration value to
allow for that if explicitly configured.

Fixes: #34
Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>