Add a retry timer to rpcJoinNode()
Prior to this commit, there was a race condition that would happen
if two nodes tried to join the cluster in close succession. This
would cause rpcJoinNode() to hang while waiting for messages from
etcd. This commit adds a retry timer to try a few times before
giving up.

In order to test this, a test had to be adjusted to be
multi-threaded so that we could join two nodes simultaneously. For
this to work, a few adjustments were needed in base.go so that
concurrency would be safe.

fixes #78

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>