The non-blocking hook tests announces itself
I've been investigating a test that occasionally fails, and I've
observed that it seems that the non-blocking hook script may be
getting executed when the blocking one is expected. This commit
adds an echo to the non-blocking hook test script so that it
announces itself, making it easier to identify in the test
output.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>