Raise the timeout on waiting for the state store
The tests ocassionally fail due to hitting a timeout while waiting
for the state store to come online. This commit raises the timeout
to increase the reliability of the tests.

While testing with the test suite on my laptop, I found that 8s
seemed to be sufficient, but I went with 32s arbitrarily, as it
seems to help Travis CI to pass more reliably.

Note that this change will affect how long production servers will
wait as well, since the timeout is in production code and not just
test code.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>