Add a file buffered chan
repospanner currently uses sqlite as a task queue to note which
objects need to get synced out to which nodes. This works
correctly, but is slow due to the number of fsyncs called during
operation due to all the commits. It is also slow due to locking
the database to ensure only one transaction happens at a time,
since sqlite is not safe for parallel access.

This commit adds a new API to set up a file buffered chan. The
API creates two chans for the caller to use to send objects in and
get objects back out. In the middle, the objects are stored in a
file to ensure that we limit memory usage. These chans are
intended to be used as task queues to replace sqlite.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>