Remove the usage of sqlite from the project
This commit removes the usage of sqlite from the project. This is
a tradeoff. It causes the project to lose the ability to track
whether each individual git object has made it to a majority of
nodes, meaning that if any node misses a single object, that node
is considered to have failed the push. In exchange, git pushes
are now significantly faster.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>