Remove a node from performHookTest()
Prior to this commit, performHookTest() would ocassionally fail
due to a race between synchronizing the new hooks to all three
nodes and asserting that the hook is in place on the nodes. The
call to ask repospanner to accept the new hook returns once the
hook reaches a majority of nodes, which in this case was two. This
means that if it asserted that the hook was in place on the third
node that didn't have it yet, it would fail.

This commit reduces the number of nodes to 2, so that a "majority"
means "all nodes", so that there is no longer a race that the
assertion can win.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>